+++ sphinx_lua_ls/__init__.py
@@ -70,6 +70,7 @@ def run_lua_ls(app: sphinx.application.Sphinx):
             max_version = "4.0.0"
         else:
             max_version = "2.0.0"
+    skip_versions = domain.config.skip_versions
 
     cwd = pathlib.Path.cwd()
     try:
@@ -77,6 +78,7 @@ def run_lua_ls(app: sphinx.application.Sphinx):
             backend=domain.config.backend,
             min_version=min_version,
             max_version=max_version,
+            skip_versions=skip_versions or [],
             cwd=root_dir,
             reporter=sphinx_lua_ls.lua_ls.SphinxProgressReporter(app.verbosity),
             install=domain.config.auto_install,
@@ -195,6 +197,7 @@ def setup(app: sphinx.application.Sphinx):
     app.add_config_value("lua_ls_auto_install_location", None, rebuild="")
     app.add_config_value("lua_ls_min_version", None, rebuild="env")
     app.add_config_value("lua_ls_max_version", "__auto__", rebuild="env")
+    app.add_config_value("lua_ls_skip_versions", None, rebuild="env")
     app.add_config_value("lua_ls_lua_version", None, rebuild="html")
     app.add_config_value("lua_ls_default_options", None, rebuild="env")
     app.add_config_value("lua_ls_apidoc_roots", None, rebuild="")